fix: exercise 5
